screen-orientation/browsing-contexts.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script>
  async function makeIFrame(src = "about:blank") {
    const iframe = document.createElement("iframe");
    document.body.appendChild(iframe);
    await new Promise(r => {
      if (iframe.contentDocument.readyState === "complete") {
        return r(); // it's loaded
      }
      iframe.onload = r;
    });
    return iframe;
  }

  promise_test(async t => {
    const iframe = await makeIFrame();
    await test_driver.bless("request full screen", () => {
      return document.documentElement.requestFullscreen();
    });
    const currentOrientation = screen.orientation.type;
    const isPortrait = currentOrientation.includes("portrait");
    const newOrientation = `${isPortrait ? "landscape" : "portrait"}`;
    const promiseToLockIFrame = iframe.contentWindow.screen.orientation.lock(
      `${newOrientation}-secondary`
    );
    await screen.orientation.lock(`${newOrientation}-primary`);
    await promise_rejects(
      t,
      "AbortError",
      promiseToLockIFrame,
      "Descendant browsing context's pending promise is rejected"
    );
    screen.orientation.unlock();
    await document.exitFullscreen();
    iframe.remove();
  }, "Browsing contexts test");

  promise_test(async t => {
    const iframe = await makeIFrame();
    const promiseToChange = new Promise(resolve => {
      screen.orientation.addEventListener("change", resolve);
    });
    await test_driver.bless("request full screen", () => {
      return document.documentElement.requestFullscreen();
    });
    const originalOrientation = screen.orientation.type;
    const isPortrait = originalOrientation.startsWith("portrait");
    const newOrientation = `${isPortrait ? "landscape" : "portrait"}-primary`;
    await iframe.contentWindow.screen.orientation.lock(newOrientation);
    window.location.href = "https://not-web-platform.test:8443";
    await promiseToChange;
    assert_equals(
      screen.orientation.type,
      originalOrientation,
      "Navigating from iframe to new top-level browsing context unlocks orientation"
    );
    screen.orientation.unlock();
    await document.exitFullscreen();
    iframe.remove();
  }, "Iframe navigation test");
</script>
